YUBICO_MODE: enroll yubikey with randomprefix=6 as default
Until now a mass enrolled yubikey would only work with pam_yubico when
we passwd --yubicorandomprefix 6. Use that as a default for newly
enrolled tokens. Old tokens will still work, but not for pam_yubico,
but that's not a regression.

The OTP generated by a yubikey in AES mode must be 44 characters for pam_yubico.
The first 12 characters are the public ID of the token, the rest will be the
OTP.  When initializing with the yubikey personalization tool you can create an
optional private identity that is included as an input parameter in the OTP
generation algorithm.  It is not used for identification of the token.

According to http://forum.yubico.com/viewtopic.php?f=16&t=2110 the first two
characters of the modhex prefix are "cc" for factory programmed tokens and "vv"
should be used for customer programmed tokens for the Yubico cloud service.  The
personalization tool will create the prefix automatically.  To have compatible
tokens we should generate a "vv" prefix, but might want to enable users to set a
custom prefix when mass enrolling tokens.