yubikey: default prefix length in MODE_YUBICO is 6, otherwise 0
In scripts/privacyidea we use a default of -1 for --yubiprefixrandom
for MODE_YUBICO.  That way we can distinguish between no length given
on the command line, where we'll use a length of 6 and a parameter 0,
which means no prefix.

Other modes for yubikeys are unchanged.

In enrollYubikey we handle len_fixed_string == -1 for MODE_YUBICO and
use 6 as a default length.  In yubi_mass_enroll we calculate the optlen,
so we'll need to handle -1 too.

After this change we can enroll UBAM tokens (MODE_YUBICO) and use them
as well as UBOM token (OATH).